+++ src/helpers/max.ts
@@ -1,4 +1,4 @@
-export const max = (arr: Float64Array) => {
+export const max = (arr: Float32Array) => {
   let max = arr[0]
   for (let i = 1; i < arr.length; i++) {
     if (arr[i] > max) {
@@ -8,7 +8,7 @@ export const max = (arr: Float64Array) => {
   return max
 }
 
-export const min = (arr: Float64Array) => {
+export const min = (arr: Float32Array) => {
   let min = arr[0]
   for (let i = 1; i < arr.length; i++) {
     if (arr[i] < min) {
